PMA-127 - ModulePolicy

+++ app/Http/Controllers/ModuleController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\Module\UpdateRequest;
 use App\Http\Resources\ModuleResource;
 use App\Http\Resources\ModuleResourceCollection;
 use App\Repositories\Contracts\ModuleRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 
 class ModuleController extends Controller
@@ -32,9 +33,12 @@ class ModuleController extends Controller
      *
      * @param IndexRequest $request
      * @return ModuleResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', Module::class);
+
         $modules = $this->moduleRepository->findBy($request->all());
 
         return new ModuleResourceCollection($modules);
@@ -45,9 +49,12 @@ class ModuleController extends Controller
      *
      * @param  StoreRequest  $request
      * @return ModuleResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', Module::class);
+
         $module = $this->moduleRepository->save($request->all());
 
         return new ModuleResource($module);
@@ -58,9 +65,12 @@ class ModuleController extends Controller
      *
      * @param Module $module
      * @return ModuleResource
+     * @throws AuthorizationException
      */
     public function show(Module $module)
     {
+        $this->authorize('show', $module);
+
         return new ModuleResource($module);
     }
 
@@ -70,9 +80,12 @@ class ModuleController extends Controller
      * @param UpdateRequest $request
      * @param Module $module
      * @return ModuleResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, Module $module)
     {
+        $this->authorize('update', $module);
+
         $module = $this->moduleRepository->update($module, $request->all());
 
         return new ModuleResource($module);
@@ -83,9 +96,12 @@ class ModuleController extends Controller
      *
      * @param Module $module
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(Module $module)
     {
+        $this->authorize('destroy', $module);
+
         $this->moduleRepository->delete($module);
 
         return response()->json(null, 204);
